+++ CHANGES.md
@@ -14,7 +14,7 @@
 * BigQuery Storage API has been upgraded to version 2.47.0
 * GAX has been upgraded to version 2.38.0
 * gRPC has been upgraded to version 1.60.0
-* Netty has been upgraded to version 4.1.103.Final
+* Netty has been upgraded to version 4.1.101.Final
 * Protocol Buffers has been upgraded to version 3.25.1
 
 ## 0.34.0 - 2023-10-31